tray sizing and the start of style improvements
