+++ src/components/LetterTray/LetterTray.css
@@ -1,5 +1,8 @@
 .tray-container{
     display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
 }
 .tile-container{
     margin: 5px;
