+++ src/components/board/Board.css
@@ -6,6 +6,12 @@
     margin-top: 30px;
 }
 
+.square{
+    height: 50px;
+    width: 50px;
+}
+
+
 .tile-divider{
     width: 25%;
     pointer-events: none;
@@ -26,32 +32,24 @@
 }
 
 .empty-square{
-    width: 50px;
-    height: 50px;
     border: 4px rgb(176, 176, 176) solid;
     margin: 2px;
     border-radius: 7px;
     background-color: lightgray;
 }
 .empty-square-final{
-    width: 50px;
-    height: 50px;
     border: 4px black solid;
     margin: 2px;
     border-radius: 7px;
     background-color: lightgray;
 }
 .rock-square{
-    width: 50px;
-    height: 50px;
     background-color: transparent;
     margin: 6px;
     border-radius: 7px;
 
 }
 .letter-square{
-    width: 50px;
-    height: 50px;
     background-color: #ffe27a;
     border: 4px #f1c52f solid;
     border-radius: 7px;
@@ -64,8 +62,6 @@
     cursor: grab;
 }
 .letter-square-final{
-    width: 50px;
-    height: 50px;
     background-color: #ffe27a;
     border: 4px black solid;
     border-radius: 7px;
@@ -78,18 +74,9 @@
     cursor: grab;
 }
 
-.start-square{
-    width: 1px;
-    height: 50px;
-    border: 3px #f03c27 solid;
-    margin: 1px;
-    border-radius: 5px;
+@media (max-width:700px) {
+    .square{
+        width: 40px;
+        height: 40px;
+    }
 }
-.empty-start-square{
-    width: 1px;
-    height: 50px;
-    border: 3px white solid;
-    margin: 1px;
-    border-radius: 5px;
-}
-
